backend/django_backend/pyproject.toml
[project]
name = "data-analyze"
version = "0.1.0"
description = ""
authors = [
    { name = "Florenz", email = "florenz@tju.edu.cn" }
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "accelerate>=1.10.1",
    "chromadb>=1.2.0",
    "django>=5.2.7",
    "django-cors-headers>=4.9.0",
    "django-ninja>=1.4.4",
    "hf-xet>=1.1.10",
    "langchain==0.3.27",
    "langchain-community>=0.3.31",
    "langchain-core>=0.3.79",
    "langchain-huggingface>=0.3.1",
    "langchain-ollama>=0.3.10",
    "langchain-openai>=0.3.35",
    "llama-index>=0.14.5",
    "llama-index-core>=0.14.5",
    "llama-index-embeddings-langchain>=0.4.1",
    "llama-index-llms-langchain>=0.7.1",
    "llama-index-vector-stores-chroma>=0.5.3",
    "ollama>=0.6.0",
    "openai>=1.109.1",
    "python-dotenv>=1.1.1",
    "sentence-transformers>=5.1.1",
    "setuptools>=80.9.0",
    "torch>=2.9.0",
    "transformers>=4.57.1",
]
